update launch script

+++ scripts/launch.py
@@ -145,18 +145,19 @@ def launch_ws_reverse_server(server_host: string, server_port: int, forwards, de
             if debug:
                 log.info('%-21s %13s %-4s %-32s %-3s %s', dt[11:], filename, lineno, fn, sessionId or '-', message)
 
-def launch_ws_reverse_client(server_host, server_port, forward, debug=False, mode='r-wsclient'):
+def launch_ws_reverse_client(server_host, server_port, forward, debug=False, mode='r-wsclient', path='/ws', exe='kptunnel', compile=True):
     log = logging.getLogger('wsc')
     log.info('%s client %s %s', '-' * 20, datetime.now().isoformat(), '-' * 20)
 
-    # subprocess.run(['rm', '-rf', 'kptunnel*'])
-    pathlib.Path('kptunnel.exe').unlink(missing_ok=True)
+    if compile:
+        # subprocess.run(['rm', '-rf', 'kptunnel*'])
+        pathlib.Path('kptunnel.exe').unlink(missing_ok=True)
 
-    subprocess.run(['go', 'build'])
-    log.info('kpt compiled')
+        subprocess.run(['go', 'build'])
+        log.info('kpt compiled')
 
     log.info('connect to %s:%s, forward: %s', server_host, server_port, forward)
-    cmd =  ['./kptunnel', 'r-wsclient', f'{server_host}:{server_port}', '-tls', '-wspath', '/ws', '-pass', 'cpass', '-encPass', 'tpass']
+    cmd =  ['./kptunnel', 'r-wsclient', f'{server_host}:{server_port}', '-tls', '-wspath', path, '-pass', 'cpass', '-encPass', 'tpass']
     with subprocess.Popen(cmd, stdout=subprocess.PIPE) as p:
         for line in p.stdout:
             j = json.loads(line.strip())
@@ -176,14 +177,26 @@ def main():
     parser.add_argument('--port', type=int, default=443)
     parser.add_argument('-f', '--forwards', type=str, nargs='*', default=[]) # collect all as a list: -f a b => [a, b]
     args = parser.parse_args()
-    if args.mode == 'client':
+    print(f'mode: {args.mode}')
+    if args.mode == 'ws-client':
         while True:
             launch_ws_reverse_client(args.host, args.port, [], debug=True)
             time.sleep(3)
-    if args.mode == 'server':
+
+    if args.mode == 'dev-client':
+        while True:
+            launch_ws_reverse_client(args.host, args.port, [], debug=True, path='/dev', exe='kpt')
+            time.sleep(3)
+
+    if args.mode == 'ws-server':
         forwards = [':2222,localhost:2222', ':2223,localhost:2223', ':2224,localhost:2224', *args.forwards]
         print(f"forwards: {forwards}")
-        launch_ws_reverse_server('127.0.0.1', args.port, forwards, debug=True, mode='r-wsserver')
+        launch_ws_reverse_server('127.0.0.1', 34022, forwards, debug=True, mode='r-wsserver')
+
+    if args.mode == 'dev-server':
+        forwards = [':4122,127.0.0.1:22', *args.forwards]
+        print(f"forwards: {forwards}")
+        launch_ws_reverse_server('127.0.0.1', 34122, forwards, debug=True, mode='r-wsserver')
 
 if __name__ == '__main__':
     main()
